requirements.txt
absl-py==0.2.2
asn1crypto==0.24.0
astor==0.6.2
attrs==17.4.0
Automat==0.6.0
bleach==1.5.0
bootstrap-admin==0.3.8
cachetools==2.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
cycler==0.10.0
deepgram-brain==0.1.1
defusedxml==0.5.0
dill==0.2.7.1
Django==1.11.8
django-allauth==0.34.0
django-bootstrap4==0.0.6
django-fa==1.0.0
django-fileprovider==0.1.4
eyeD3==0.8.5
Flask==1.0.2
Flask-Cors==3.0.6
fontawesome==4.7.0.post5
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gast==0.2.0
google-api-core==1.1.1
google-auth==1.4.1
google-cloud==0.32.0
google-cloud-bigquery==0.28.0
google-cloud-bigquery-datatransfer==0.1.0
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.0
google-cloud-core==0.28.1
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.0
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.0
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.0
google-cloud-runtimeconfig==0.28.0
google-cloud-spanner==0.29.0
google-cloud-speech==0.33.0
google-cloud-storage==1.8.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.0
google-cloud-videointelligence==1.0.0
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.12.0
h5py==2.8.0
hmmlearn==0.2.0
html5lib==0.9999999
httplib2==0.10.3
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.10
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
language-check==1.1
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
nltk==3.2.5
numpy==1.14.3
oauth2client==3.0.0
oauthlib==2.0.6
ontospy==1.9.1
opencv-contrib-python==3.4.1.15
Owlready2==0.5
parsel==1.4.0
pathlib==1.0.1
Pillow==5.1.0
ply==3.8
pronto==0.10.2
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.5.2.post1
psutil==5.4.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pyAudioAnalysis==0.1.3
pycparser==2.18
PyDispatcher==2.0.5
pydub==0.21.0
pyfiglet==0.7.5
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-magic==0.4.15
python3-openid==3.1.0
pytz==2018.4
PyYAML==3.12
queuelib==1.5.0
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
scikit-learn==0.19.1
scipy==1.0.1
Scrapy==1.5.0
service-identity==17.0.0
simplejson==3.15.0
six==1.11.0
sklearn==0.0
SPARQLWrapper==1.8.2
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
Twisted==17.9.0
urllib3==1.22
w3lib==1.19.0
Wave==0.0.2
Werkzeug==0.14.1
zope.interface==4.5.0
